consumer/requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.3.9
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
kafka-python==1.4.6
multidict==4.5.2
pymongo==3.8.0
requests==2.22.0
typing-extensions==3.7.2
urllib3==1.25.2
yarl==1.3.0
